Fixed File Explorer focus inconsistencies and moved (most of) the File Explorer click handling to Rust

+++ core/src/commands/file_explorer/types.rs
@@ -171,7 +171,7 @@ impl FromStr for FileExplorerCommand {
 }
 
 pub struct FileExplorerContext {
-    pub index: usize,
+    pub index: i64,
     pub item_path: PathBuf,
     pub target_is_item: bool,
     pub item_is_directory: bool,
